pyproject.toml
[project]
name = "radar-basics"
version = "0.1.0"
description = "Educational pulsed phased-array radar echo simulator"
authors = [
    { name = "Yan", email = "yanzhanggc@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "pytest>=9.0.3",
]
